app/requirements/dev.txt
autopep8>=1.5.4
mypy>=0.790
pycodestyle>=2.6.0
pydocstyle>=5.1.1
pylint>=2.6.0
pytest>=6.2.1
pika-stubs>=0.1.3
